Moved StreamHelper to most basic usage
