+++ FileHelpers/Engines/FileHelperEngine.cs
@@ -6,6 +6,7 @@ using System.Data;
 using System.Diagnostics;
 using System.IO;
 using System.Text;
+using FileHelpers.Engines;
 using FileHelpers.Events;
 using FileHelpers.Helpers;
 using FileHelpers.Streams;
